Add paper type definitions from escpr2
Epson provides a separate driver, named escpr2, which provides support
for a limited set of printers. escpr2 is not fully open-source.

With an Epson printer the paper setting in the printer menu, must match
the print setting. Certain paper types, such as "Glossy" are not
understood by vanilla escpr, so such settings cannot be used. Although
"Premium Glossy" could be used instead, this is not ideal for certain
photo papers, as the ink density and color profile differ slightly.

To enable all paper types, import the extra types from escpr2.